Added user auth

+++ admin/page/delete.php
@@ -2,14 +2,18 @@
 
 require_once "../../head.php";
 
-$id = Router::get('id');
+if ($User->isLogged()) { 
+	$id = Router::get('id');
 
-if ($id) {
-	
-	$Page->delete($id);
+	if ($id) {
+		
+		$Page->delete($id);
 
-	Server::redirect('../pages.php');
+		Server::redirect('../pages.php');
 
+	} else {
+		// Delete Error
+	}
 } else {
-	// Delete Error
+	Server::redirect('/login');
 }
