+++ admin/project/delete.php
@@ -2,18 +2,22 @@
 
 require_once "../../head.php";
 
-// Get project id
-$id = Router::get('id');
+if ($User->isLogged()) { 
+	// Get project id
+	$id = Router::get('id');
 
-// There is id set?
-if ($id) {
-	
-	// Delete targeted project
-	$Project->delete($id);
+	// There is id set?
+	if ($id) {
+		
+		// Delete targeted project
+		$Project->delete($id);
 
-	// Redirect to project list
-	Server::redirect('../projects.php');
+		// Redirect to project list
+		Server::redirect('../projects.php');
 
+	} else {
+		// Delete Error
+	}
 } else {
-	// Delete Error
+	Server::redirect('/login');
 }
